safari-534.53-branch/LayoutTests/editing/execCommand/5049671.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div id="div" contenteditable="true"><a href="http://www.google.com/">This shouldn't be a link or underlined.</a></div>
<script>

Markup.description('This tests was added for a bug where Remove Format would fail to remove links that were fully selected.'
 + ' However, because the bug 43017 requires WebKit does not remove anchor elements, RemoveFormat should NOT remove anchor elements.');

var div = document.getElementById("div");
div.focus();
document.execCommand("SelectAll");

Markup.dump(div, 'Before RemoveFormat');

document.execCommand("RemoveFormat");

Markup.dump(div, 'After RemoveFormat');

</script>
</body>
</html>
