third_party/blink/web_tests/platform/generic/editing/execCommand/5049671-expected.txt

This tests was added for a bug where Remove Format would fail to remove links that were fully selected. However, because the bug 43017 requires WebKit does not remove anchor elements, RemoveFormat should NOT remove anchor elements.

Before RemoveFormat:
| <a>
|   href="http://www.google.com/"
|   "<#selection-anchor>This shouldn't be a link or underlined.<#selection-focus>"

After RemoveFormat:
| <a>
|   href="http://www.google.com/"
|   "<#selection-anchor>This shouldn't be a link or underlined.<#selection-focus>"
